[c_book] fix a bug in 1.4 example - 0 fahr was not printing if int

+++ c_book/1-ch/ex-1-5.c
@@ -6,10 +6,10 @@
 
 int main(void)
 {
-  int fahr;
+  float fahr;
 
   for (fahr = LOWER; fahr <= UPPER; fahr += STEP)
-    printf("%3.0d %6.1f\n", fahr, (5.0/9.0) * (fahr - 32)); 
+    printf("%3.0f %6.1f\n", fahr, (5.0/9.0) * (fahr - 32.0)); 
 
   return 0;
 }
